Imported Upstream version 1.9.3.125

+++ test/win32ole/test_err_in_callback.rb
@@ -7,7 +7,6 @@ begin
   require 'win32ole'
 rescue LoadError
 end
-require 'rbconfig'
 if defined?(WIN32OLE)
   require 'mkmf'
   require 'test/unit'
@@ -20,11 +19,21 @@ if defined?(WIN32OLE)
         @iopt = $:.map {|e|
           " -I " + e
         }.join("")
-        @script = File.dirname(__FILE__) + "/err_in_callback.rb"
+        @script = File.join(File.dirname(__FILE__), "err_in_callback.rb")
       end
     end
 
+    def available_adodb?
+      begin
+        db = WIN32OLE.new('ADODB.Connection')
+      rescue WIN32OLERuntimeError
+        return false
+      end
+      return true
+    end
+
     def test_err_in_callback
+      skip "'ADODB.Connection' is not available" unless available_adodb?
       if @ruby
         cmd = "#{@ruby} -v #{@iopt} #{@script} > test_err_in_callback.log 2>&1"
         system(cmd)
@@ -37,8 +46,7 @@ if defined?(WIN32OLE)
     end
 
     def teardown
-      File.unlink("test_err_in_callback.log")
+      File.unlink("test_err_in_callback.log") if File.exist?("test_err_in_callback.log")
     end
-
   end
 end
